+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 2.6.4
 
-- Fixes issue where markdown was rendered as plain text on scorecard service page
+- Fixes issue where markdown was rendered as plain text on Scorecard service page
 
 ### 2.6.3